fix(#73): 사진 여러 장에서 한 장 업로드로 기능 수정
