+++ src/main/java/com/neo/neomarket/service/impl/S3FileUploadService.java
@@ -1,0 +1,42 @@
+package com.neo.neomarket.service.impl;
+
+import com.amazonaws.AmazonServiceException;
+import com.amazonaws.services.s3.AmazonS3Client;
+import com.amazonaws.services.s3.model.CannedAccessControlList;
+import com.amazonaws.services.s3.model.ObjectMetadata;
+import com.amazonaws.services.s3.model.PutObjectRequest;
+import java.io.IOException;
+import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+
+@RequiredArgsConstructor
+@Service
+public class S3FileUploadService {
+
+    private final AmazonS3Client amazonS3Client;
+
+    @Value("${cloud.aws.s3.bucketName}")
+    private String bucketName;
+
+    public String uploadFile(MultipartFile file) throws IOException {
+        try {
+            String fileName = generateFileName(file);
+            ObjectMetadata metadata = new ObjectMetadata();
+            metadata.setContentLength(file.getSize());
+            amazonS3Client.putObject(
+                    new PutObjectRequest(bucketName, fileName, file.getInputStream(), metadata).withCannedAcl(
+                            CannedAccessControlList.PublicRead));
+            return amazonS3Client.getUrl(bucketName, fileName).toString();
+        } catch (AmazonServiceException e) {
+            throw new IOException("Error uploading file to S3", e);
+        }
+    }
+
+    private String generateFileName(MultipartFile file) {
+        return UUID.randomUUID().toString() + "-" + file.getOriginalFilename().replace(" ", "_");
+    }
+
+}
